fix: remove Vault — switch to vanilla item economy (V3.3.1)
- Replaced Vault economy with VanillaEconomy (physical items as currency)
- Default currency: Diamond (configurable via economy.currency-item)
- Prices are integer item counts, e.g. "/ah sell 5" = 5 diamonds
- Seller paid directly if online, via delivery if offline
- Builder shop purchases now use item economy
- Currency item cannot be listed in auction house
- Removed Vault repo and VaultAPI dependency from pom.xml
- Removed softdepend: Vault from plugin.yml
- Updated all GUI displays to show "5x Diamond" format
- Updated lang files: not-enough-items, purchase-success messages
- Removed listing-fee-percent (was Vault-only feature)
- Separated README: Vietnamese + English, clean and concise
- JAR moved to GitHub Releases (not committed to repo)

+++ src/main/java/dev/zerep/zeah/commands/AHCommand.java
@@ -46,9 +46,9 @@ public class AHCommand implements CommandExecutor, TabCompleter {
                 if (args.length < 2) {
                     player.sendMessage(plugin.getLang().format("auction.usage")); return true;
                 }
-                double price;
+                int price;
                 try {
-                    price = Double.parseDouble(args[1]);
+                    price = Integer.parseInt(args[1]);
                     if (price <= 0) throw new NumberFormatException();
                 } catch (NumberFormatException e) {
                     player.sendMessage(plugin.getLang().format("invalid-number", "input", args[1])); return true;
